requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.3
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-ftp==3.6.0
apache-airflow-providers-http==4.6.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-postgres==5.11.3
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
argcomplete==3.1.3
asgiref==3.7.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backoff==1.10.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.2.2
ConfigUpdater==3.1.1
connexion==2.14.2
cron_descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.5
decorator==5.2.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docker==7.1.0
docutils==0.20.1
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
great-expectations==1.1.0
greenlet==3.0.1
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyterlab_widgets==3.0.13
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
makefun==1.15.6
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.1.2
monotonic==1.6
more-itertools==10.5.0
msgspec==0.18.6
multidict==6.0.4
numpy==1.24.4
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
packaging==23.2
pandas==2.0.3
parso==0.8.4
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.9.0
pickleshare==0.7.5
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
pluggy==1.3.0
posthog==2.5.0
prison==0.2.1
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==4.24.4
psutil==5.9.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.4
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.4.0
rpds-py==0.10.6
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
scipy==1.10.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.8.0
tzdata==2025.1
tzlocal==5.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==2.2.3
widgetsnbextension==4.0.13
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0